Improve /patch functionality Fixes #3
